requirements.txt
torch==2.0.1
transformers==4.34.0
huggingface-hub==0.17.3
accelerate==0.25.0
bitsandbytes==0.41.1
peft==0.5.0
open-clip-torch==2.20.0
openai-whisper==20231117
gradio==4.1.1
Pillow==10.0.0
numpy>=1.24.3
tqdm>=4.65.0
soundfile>=0.12.1
datasets